+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- New version information is automatically added after line 8 -->
 
-## Version 1.7.13 - tbd
+## Version 1.7.13 - 2021-10-18
 
 ### Fixed
+
+- Escape quotes in search string before quoting
 
 ## Version 1.7.12 - 2021-10-07